cv-service/requirements.txt
# flask==3.1.0
# flask-cors==5.0.1
# torch==2.2.0
# torchvision==0.17.0
# Pillow==10.3.0
# numpy==2.2.3
# gunicorn==23.0.0
numpy==1.26.4
flask==3.1.0
flask-cors==5.0.1
torch==2.2.0
torchvision==0.17.0
Pillow==10.3.0
gunicorn==23.0.0